+++ dash-renderer/src/actions/api.js
@@ -1,4 +1,3 @@
-/* global fetch: true */
 import {mergeDeepRight} from 'ramda';
 import {handleAsyncError, getCSRFHeader} from '../actions';
 import {urlBase} from '../utils';