+++ dash/dash.py
@@ -221,7 +221,7 @@ class Dash(object):
         requests_pathname_prefix=None,
         routes_pathname_prefix=None,
         serve_locally=True,
-        compress=True,
+        compress=None,
         meta_tags=None,
         index_string=_default_index,
         external_scripts=None,
@@ -238,9 +238,7 @@ class Dash(object):
                     "since v1.0. See https://dash.plotly.com for details."
                 )
             # any other kwarg mimic the built-in exception
-            raise TypeError(
-                "Dash() got an unexpected keyword argument '" + key + "'"
-            )
+            raise TypeError("Dash() got an unexpected keyword argument '" + key + "'")
 
         # We have 3 cases: server is either True (we create the server), False
         # (defer server creation) or a Flask app instance (we use their server)
@@ -276,14 +274,12 @@ class Dash(object):
             routes_pathname_prefix=routes_prefix,
             requests_pathname_prefix=requests_prefix,
             serve_locally=serve_locally,
-            compress=compress,
+            compress=get_combined_config("compress", compress, True),
             meta_tags=meta_tags or [],
             external_scripts=external_scripts or [],
             external_stylesheets=external_stylesheets or [],
             suppress_callback_exceptions=get_combined_config(
-                "suppress_callback_exceptions",
-                suppress_callback_exceptions,
-                False,
+                "suppress_callback_exceptions", suppress_callback_exceptions, False,
             ),
             show_undo_redo=show_undo_redo,
         )
@@ -420,9 +416,7 @@ class Dash(object):
         # catch-all for front-end routes, used by dcc.Location
         self._add_url("{}<path:path>".format(prefix), self.index)
 
-        self._add_url(
-            "{}_favicon.ico".format(prefix), self._serve_default_favicon
-        )
+        self._add_url("{}_favicon.ico".format(prefix), self._serve_default_favicon)
 
     def _add_url(self, name, view_func, methods=("GET",)):
         self.server.add_url_rule(
@@ -536,8 +530,7 @@ class Dash(object):
         def _relative_url_path(relative_package_path="", namespace=""):
 
             module_path = os.path.join(
-                os.path.dirname(sys.modules[namespace].__file__),
-                relative_package_path,
+                os.path.dirname(sys.modules[namespace].__file__), relative_package_path,
             )
 
             modified = int(os.stat(module_path).st_mtime)
@@ -577,9 +570,7 @@ class Dash(object):
                     else:
                         srcs += resource["external_url"]
             elif "absolute_path" in resource:
-                raise Exception(
-                    "Serving files from absolute_path isn't supported yet"
-                )
+                raise Exception("Serving files from absolute_path isn't supported yet")
             elif "asset_path" in resource:
                 static_url = self.get_asset_url(resource["asset_path"])
                 # Add a cache-busting query param
@@ -622,9 +613,7 @@ class Dash(object):
         dev = self._dev_tools.serve_dev_bundles
         srcs = (
             self._collect_and_register_resources(
-                self.scripts._resources._filter_resources(
-                    deps, dev_bundles=dev
-                )
+                self.scripts._resources._filter_resources(deps, dev_bundles=dev)
             )
             + self.config.external_scripts
             + self._collect_and_register_resources(
@@ -641,18 +630,13 @@ class Dash(object):
                 if isinstance(src, dict)
                 else '<script src="{}"></script>'.format(src)
                 for src in srcs
-            ] +
-            [
-                '<script>{}</script>'.format(src)
-                for src in self._inline_scripts
             ]
+            + ["<script>{}</script>".format(src) for src in self._inline_scripts]
         )
 
     def _generate_config_html(self):
         return (
-            '<script id="_dash-config" type="application/json">'
-            "{}"
-            "</script>"
+            '<script id="_dash-config" type="application/json">{}</script>'
         ).format(json.dumps(self._config()))
 
     def _generate_renderer(self):
@@ -671,9 +655,7 @@ class Dash(object):
 
         tags = []
         if not has_ie_compat:
-            tags.append(
-                '<meta http-equiv="X-UA-Compatible" content="IE=edge">'
-            )
+            tags.append('<meta http-equiv="X-UA-Compatible" content="IE=edge">')
         if not has_charset:
             tags.append('<meta charset="UTF-8">')
 
@@ -683,9 +665,7 @@ class Dash(object):
 
     # Serve the JS bundles for each package
     def serve_component_suites(self, package_name, path_in_package_dist):
-        path_in_package_dist, has_fingerprint = check_fingerprint(
-            path_in_package_dist
-        )
+        path_in_package_dist, has_fingerprint = check_fingerprint(path_in_package_dist)
 
         if package_name not in self.registered_paths:
             raise exceptions.DependencyException(
@@ -723,8 +703,7 @@ class Dash(object):
         )
 
         response = flask.Response(
-            pkgutil.get_data(package_name, path_in_package_dist),
-            mimetype=mimetype,
+            pkgutil.get_data(package_name, path_in_package_dist), mimetype=mimetype,
         )
 
         if has_fingerprint:
@@ -737,7 +716,7 @@ class Dash(object):
             response.add_etag()
             tag = response.get_etag()[0]
 
-            request_etag = flask.request.headers.get('If-None-Match')
+            request_etag = flask.request.headers.get("If-None-Match")
 
             if '"{}"'.format(tag) == request_etag:
                 response = flask.Response(None, status=304)
@@ -909,9 +888,7 @@ class Dash(object):
                 if not isinstance(arg, obj):
                     raise exceptions.IncorrectTypeException(
                         "The {} argument `{}` must be "
-                        "of type `dash.{}`.".format(
-                            name.lower(), str(arg), name
-                        )
+                        "of type `dash.{}`.".format(name.lower(), str(arg), name)
                     )
 
                 invalid_characters = ["."]
@@ -948,9 +925,7 @@ class Dash(object):
                             ).format(arg_id, all_ids)
                         )
 
-                    component = (
-                        layout if layout_id == arg_id else layout[arg_id]
-                    )
+                    component = layout if layout_id == arg_id else layout[arg_id]
 
                     if (
                         arg_prop
@@ -970,9 +945,7 @@ class Dash(object):
                             {2}
                         """
                             ).format(
-                                arg_prop,
-                                arg_id,
-                                component.available_properties,
+                                arg_prop, arg_id, component.available_properties,
                             )
                         )
 
@@ -997,9 +970,7 @@ class Dash(object):
                 (Subscribing to input components will cause the
                 callback to be called whenever their values change.)
             """
-                ).format(
-                    len(state), "elements" if len(state) > 1 else "element"
-                )
+                ).format(len(state), "elements" if len(state) > 1 else "element")
             )
 
         for i in inputs:
@@ -1023,9 +994,7 @@ class Dash(object):
                     "multi output callback!\n Duplicates:\n {}".format(
                         ",\n".join(
                             k
-                            for k, v in (
-                                (str(x), output.count(x)) for x in output
-                            )
+                            for k, v in ((str(x), output.count(x)) for x in output)
                             if v > 1
                         )
                     )
@@ -1045,9 +1014,7 @@ class Dash(object):
         if is_multi:
 
             def duplicate_check():
-                ns["duplicates"] = callbacks.intersection(
-                    str(y) for y in output
-                )
+                ns["duplicates"] = callbacks.intersection(str(y) for y in output)
                 return ns["duplicates"]
 
         else:
@@ -1080,9 +1047,7 @@ class Dash(object):
     def _validate_callback_output(output_value, output):
         valid = [str, dict, int, float, type(None), Component]
 
-        def _raise_invalid(
-            bad_val, outer_val, path, index=None, toplevel=False
-        ):
+        def _raise_invalid(bad_val, outer_val, path, index=None, toplevel=False):
             bad_type = type(bad_val).__name__
             outer_id = (
                 "(id={:s})".format(outer_val.id)
@@ -1119,9 +1084,7 @@ class Dash(object):
                     location=(
                         "\n"
                         + (
-                            "[{:d}] {:s} {:s}".format(
-                                index, outer_type, outer_id
-                            )
+                            "[{:d}] {:s} {:s}".format(index, outer_type, outer_id)
                             if index is not None
                             else ("[*] " + outer_type + " " + outer_id)
                         )
@@ -1149,16 +1112,12 @@ class Dash(object):
                 for p, j in val._traverse_with_paths():
                     # check each component value in the tree
                     if not _value_is_valid(j):
-                        _raise_invalid(
-                            bad_val=j, outer_val=val, path=p, index=index
-                        )
+                        _raise_invalid(bad_val=j, outer_val=val, path=p, index=index)
 
                     # Children that are not of type Component or
                     # list/tuple not returned by traverse
                     child = getattr(j, "children", None)
-                    if not isinstance(
-                        child, (tuple, collections.MutableSequence)
-                    ):
+                    if not isinstance(child, (tuple, collections.MutableSequence)):
                         if child and not _value_is_valid(child):
                             _raise_invalid(
                                 bad_val=child,
@@ -1196,9 +1155,7 @@ class Dash(object):
             _validate_value(output_value)
 
     # pylint: disable=dangerous-default-value
-    def clientside_callback(
-        self, clientside_function, output, inputs=[], state=[]
-    ):
+    def clientside_callback(self, clientside_function, output, inputs=[], state=[]):
         """Create a callback that updates the output by calling a clientside
         (JavaScript) function instead of a Python function.
 
@@ -1270,17 +1227,19 @@ class Dash(object):
             if isinstance(output, (list, tuple)):
                 out0 = output[0]
 
-            namespace = '_dashprivate_{}'.format(out0.component_id)
-            function_name = '{}'.format(out0.component_property)
+            namespace = "_dashprivate_{}".format(out0.component_id)
+            function_name = "{}".format(out0.component_property)
 
             self._inline_scripts.append(
                 """
                 var clientside = window.dash_clientside = window.dash_clientside || {{}};
                 var ns = clientside["{0}"] = clientside["{0}"] || {{}};
                 ns["{1}"] = {2};
-                """.format(namespace.replace('"', '\\"'),
-                           function_name.replace('"', '\\"'),
-                           clientside_function)
+                """.format(
+                    namespace.replace('"', '\\"'),
+                    function_name.replace('"', '\\"'),
+                    clientside_function,
+                )
             )
 
         # Callback is stored in an external asset.
@@ -1290,12 +1249,10 @@ class Dash(object):
 
         self.callback_map[callback_id] = {
             "inputs": [
-                {"id": c.component_id, "property": c.component_property}
-                for c in inputs
+                {"id": c.component_id, "property": c.component_property} for c in inputs
             ],
             "state": [
-                {"id": c.component_id, "property": c.component_property}
-                for c in state
+                {"id": c.component_id, "property": c.component_property} for c in state
             ],
             "clientside_function": {
                 "namespace": namespace,
@@ -1323,12 +1280,10 @@ class Dash(object):
 
         self.callback_map[callback_id] = {
             "inputs": [
-                {"id": c.component_id, "property": c.component_property}
-                for c in inputs
+                {"id": c.component_id, "property": c.component_property} for c in inputs
             ],
             "state": [
-                {"id": c.component_id, "property": c.component_property}
-                for c in state
+                {"id": c.component_id, "property": c.component_property} for c in state
             ],
         }
 
@@ -1373,9 +1328,7 @@ class Dash(object):
                         raise exceptions.PreventUpdate
 
                     response = {
-                        "response": {
-                            "props": {output.component_property: output_value}
-                        }
+                        "response": {"props": {output.component_property: output_value}}
                     }
 
                 try:
@@ -1396,8 +1349,7 @@ class Dash(object):
                     or lists of those.
                     """
                         ).format(
-                            property=output.component_property,
-                            id=output.component_id,
+                            property=output.component_property, id=output.component_id,
                         )
                     )
 
@@ -1418,12 +1370,10 @@ class Dash(object):
         args = []
 
         flask.g.input_values = input_values = {
-            "{}.{}".format(x["id"], x["property"]): x.get("value")
-            for x in inputs
+            "{}.{}".format(x["id"], x["property"]): x.get("value") for x in inputs
         }
         flask.g.state_values = {
-            "{}.{}".format(x["id"], x["property"]): x.get("value")
-            for x in state
+            "{}.{}".format(x["id"], x["property"]): x.get("value") for x in state
         }
         changed_props = body.get("changedPropIds")
         flask.g.triggered_inputs = (
@@ -1432,9 +1382,7 @@ class Dash(object):
             else []
         )
 
-        response = flask.g.dash_response = flask.Response(
-            mimetype="application/json"
-        )
+        response = flask.g.dash_response = flask.Response(mimetype="application/json")
 
         for component_registration in self.callback_map[output]["inputs"]:
             args.append(
@@ -1488,7 +1436,7 @@ class Dash(object):
         eager_loading = self.config.eager_loading
         for module_name in ComponentRegistry.registry:
             module = sys.modules[module_name]
-            eager = getattr(module, '_force_eager_loading', False)
+            eager = getattr(module, "_force_eager_loading", False)
             eager_loading = eager_loading or eager
 
         # Update eager_loading settings
@@ -1539,9 +1487,7 @@ class Dash(object):
                 full = os.path.join(current, f)
 
                 if f.endswith("js"):
-                    self.scripts.append_script(
-                        self._add_assets_resource(path, full)
-                    )
+                    self.scripts.append_script(self._add_assets_resource(path, full))
                 elif f.endswith("css"):
                     self.css.append_css(self._add_assets_resource(path, full))
                 elif f == "favicon.ico":
@@ -1554,8 +1500,7 @@ class Dash(object):
     @staticmethod
     def _serve_default_favicon():
         return flask.Response(
-            pkgutil.get_data("dash", "favicon.ico"),
-            content_type="image/x-icon",
+            pkgutil.get_data("dash", "favicon.ico"), content_type="image/x-icon",
         )
 
     def get_asset_url(self, path):
@@ -1604,10 +1549,7 @@ class Dash(object):
                 return chapters.page_2
         ```
         """
-        asset = _get_relative_path(
-            self.config.requests_pathname_prefix,
-            path,
-        )
+        asset = _get_relative_path(self.config.requests_pathname_prefix, path,)
 
         return asset
 
@@ -1658,10 +1600,7 @@ class Dash(object):
         `page-1/sub-page-1`
         ```
         """
-        return _strip_relative_path(
-            self.config.requests_pathname_prefix,
-            path,
-        )
+        return _strip_relative_path(self.config.requests_pathname_prefix, path,)
 
     def _setup_dev_tools(self, **kwargs):
         debug = kwargs.get("debug", False)
@@ -1685,9 +1624,7 @@ class Dash(object):
             ("hot_reload_max_retry", int, 8),
         ):
             dev_tools[attr] = _type(
-                get_combined_config(
-                    attr, kwargs.get(attr, None), default=default
-                )
+                get_combined_config(attr, kwargs.get(attr, None), default=default)
             )
 
         return dev_tools
@@ -1806,8 +1743,7 @@ class Dash(object):
                 else package.filename
                 for package in (
                     pkgutil.find_loader(x)
-                    for x in list(ComponentRegistry.registry)
-                    + ["dash_renderer"]
+                    for x in list(ComponentRegistry.registry) + ["dash_renderer"]
                 )
             ]
 
@@ -1862,9 +1798,7 @@ class Dash(object):
                 asset_path = (
                     os.path.relpath(
                         filename,
-                        os.path.commonprefix(
-                            [self.config.assets_folder, filename]
-                        ),
+                        os.path.commonprefix([self.config.assets_folder, filename]),
                     )
                     .replace("\\", "/")
                     .lstrip("/")
@@ -1907,7 +1841,8 @@ class Dash(object):
 
     def run_server(
         self,
-        port=8050,
+        host=os.getenv("HOST", "127.0.0.1"),
+        port=os.getenv("PORT", "8050"),
         debug=False,
         dev_tools_ui=None,
         dev_tools_props_check=None,
@@ -1926,7 +1861,12 @@ class Dash(object):
         If a parameter can be set by an environment variable, that is listed
         too. Values provided here take precedence over environment variables.
 
+        :param host: Host IP used to serve the application
+            env: ``HOST``
+        :type host: string
+
         :param port: Port used to serve the application
+            env: ``PORT``
         :type port: int
 
         :param debug: Set Flask debug mode and enable dev tools.
@@ -1999,9 +1939,18 @@ class Dash(object):
             dev_tools_prune_errors,
         )
 
+        # Verify port value
+        try:
+            port = int(port)
+            assert port in range(1, 65536)
+        except Exception as e:
+            e.args = [
+                "Expecting an integer from 1 to 65535, found port={}".format(repr(port))
+            ]
+            raise
+
         if self._dev_tools.silence_routes_logging:
             # Since it's silenced, the address doesn't show anymore.
-            host = flask_run_options.get("host", "127.0.0.1")
             ssl_context = flask_run_options.get("ssl_context")
             self.logger.info(
                 "Running on %s://%s:%s%s",
@@ -2021,4 +1970,4 @@ class Dash(object):
 
             self.logger.info("Debugger PIN: %s", debugger_pin)
 
-        self.server.run(port=port, debug=debug, **flask_run_options)
+        self.server.run(host=host, port=port, debug=debug, **flask_run_options)