+++ tests/integration/test_integration.py
@@ -1,11 +1,13 @@
-from multiprocessing import Value
 import datetime
 import time
-import pytest
+
 from copy import copy
+from multiprocessing import Value
+from selenium.webdriver.common.keys import Keys
+
+import pytest
 
 from bs4 import BeautifulSoup
-from selenium.webdriver.common.keys import Keys
 
 import dash_dangerously_set_inner_html
 import dash_flow_example
@@ -76,10 +78,7 @@ def test_inin002_wildcard_callback(dash_duo):
                         "string",
                         html.Div(
                             id="output-1",
-                            **{
-                                "data-cb": "initial value",
-                                "aria-cb": "initial value",
-                            }
+                            **{"data-cb": "initial value", "aria-cb": "initial value"}
                         ),
                     ]
                 )
@@ -94,9 +93,7 @@ def test_inin002_wildcard_callback(dash_duo):
         input_call_count.value += 1
         return value
 
-    @app.callback(
-        Output("output-1", "children"), [Input("output-1", "data-cb")]
-    )
+    @app.callback(Output("output-1", "children"), [Input("output-1", "data-cb")])
     def update_text(data):
         return data
 
@@ -145,9 +142,7 @@ def test_inin003_aborted_callback(dash_duo):
         raise PreventUpdate("testing callback does not update")
         return value
 
-    @app.callback(
-        Output("output2", "children"), [Input("output1", "children")]
-    )
+    @app.callback(Output("output2", "children"), [Input("output1", "children")])
     def callback2(value):
         callback2_count.value += 1
         return value
@@ -199,9 +194,7 @@ def test_inin004_wildcard_data_attributes(dash_duo):
     actual = BeautifulSoup(div.get_attribute("innerHTML"), "lxml").decode()
     expected = BeautifulSoup(
         "<div "
-        + " ".join(
-            '{}="{!s}"'.format(k, v) for k, v in attrs.items() if v is not None
-        )
+        + " ".join('{}="{!s}"'.format(k, v) for k, v in attrs.items() if v is not None)
         + "></div>",
         "lxml",
     ).decode()
@@ -246,8 +239,7 @@ def test_inin006_flow_component(dash_duo):
     )
 
     @app.callback(
-        Output("output", "children"),
-        [Input("react", "value"), Input("flow", "value")],
+        Output("output", "children"), [Input("react", "value"), Input("flow", "value")],
     )
     def display_output(react_value, flow_value):
         return html.Div(
@@ -392,15 +384,9 @@ def test_inin011_multi_output(dash_duo):
             html.Button("OUTPUT", id="output-btn"),
             html.Table(
                 [
-                    html.Thead(
-                        [html.Tr([html.Th("Output 1"), html.Th("Output 2")])]
-                    ),
+                    html.Thead([html.Tr([html.Th("Output 1"), html.Th("Output 2")])]),
                     html.Tbody(
-                        [
-                            html.Tr(
-                                [html.Td(id="output1"), html.Td(id="output2")]
-                            )
-                        ]
+                        [html.Tr([html.Td(id="output1"), html.Td(id="output2")])]
                     ),
                 ]
             ),
@@ -422,9 +408,7 @@ def test_inin011_multi_output(dash_duo):
         return n_clicks, n_clicks_timestamp
 
     # Dummy callback for DuplicateCallbackOutput test.
-    @app.callback(
-        Output("output3", "children"), [Input("output-btn", "n_clicks")]
-    )
+    @app.callback(Output("output3", "children"), [Input("output-btn", "n_clicks")])
     def dummy_callback(n_clicks):
         if n_clicks is None:
             raise PreventUpdate
@@ -433,9 +417,7 @@ def test_inin011_multi_output(dash_duo):
 
     with pytest.raises(DuplicateCallbackOutput) as err:
 
-        @app.callback(
-            Output("output1", "children"), [Input("output-btn", "n_clicks")]
-        )
+        @app.callback(Output("output1", "children"), [Input("output-btn", "n_clicks")])
         def on_click_duplicate(n_clicks):
             if n_clicks is None:
                 raise PreventUpdate
@@ -482,9 +464,7 @@ def test_inin011_multi_output(dash_duo):
         def overlapping_multi_output(n_clicks):
             return n_clicks
 
-        pytest.fail(
-            "no part of an existing multi-output can be used in another"
-        )
+        pytest.fail("no part of an existing multi-output can be used in another")
     assert (
         "{'output1.children'}" in err.value.args[0]
         or "set(['output1.children'])" in err.value.args[0]
@@ -516,11 +496,7 @@ def test_inin012_multi_output_no_update(dash_duo):
     )
 
     @app.callback(
-        [
-            Output("n1", "children"),
-            Output("n2", "children"),
-            Output("n3", "children"),
-        ],
+        [Output("n1", "children"), Output("n2", "children"), Output("n3", "children")],
         [Input("btn", "n_clicks")],
     )
     def show_clicks(n):
@@ -715,10 +691,12 @@ def test_inin015_with_custom_renderer_interpolated(dash_duo):
                     {renderer}
                     <div id="custom-footer">My custom footer</div>
                 </body>
-            </html>""".format(app_entry=kwargs["app_entry"],
-                              config=kwargs["config"],
-                              scripts=kwargs["scripts"],
-                              renderer=renderer)
+            </html>""".format(
+                app_entry=kwargs["app_entry"],
+                config=kwargs["config"],
+                scripts=kwargs["scripts"],
+                renderer=renderer,
+            )
 
     app = CustomDash()
 
@@ -757,15 +735,11 @@ def test_inin015_with_custom_renderer_interpolated(dash_duo):
 
 def test_inin016_modified_response(dash_duo):
     app = Dash(__name__)
-    app.layout = html.Div(
-        [dcc.Input(id="input", value="ab"), html.Div(id="output")]
-    )
+    app.layout = html.Div([dcc.Input(id="input", value="ab"), html.Div(id="output")])
 
     @app.callback(Output("output", "children"), [Input("input", "value")])
     def update_output(value):
-        callback_context.response.set_cookie(
-            "dash cookie", value + " - cookie"
-        )
+        callback_context.response.set_cookie("dash cookie", value + " - cookie")
         return value + " - output"
 
     dash_duo.start_server(app)
@@ -786,15 +760,10 @@ def test_inin017_late_component_register(dash_duo):
     app = Dash()
 
     app.layout = html.Div(
-        [
-            html.Button("Click me to put a dcc ", id="btn-insert"),
-            html.Div(id="output"),
-        ]
+        [html.Button("Click me to put a dcc ", id="btn-insert"), html.Div(id="output")]
     )
 
-    @app.callback(
-        Output("output", "children"), [Input("btn-insert", "n_clicks")]
-    )
+    @app.callback(Output("output", "children"), [Input("btn-insert", "n_clicks")])
     def update_output(value):
         if value is None:
             raise PreventUpdate
@@ -811,15 +780,12 @@ def test_inin017_late_component_register(dash_duo):
 
 def test_inin018_output_input_invalid_callback():
     app = Dash(__name__)
-    app.layout = html.Div(
-        [html.Div("child", id="input-output"), html.Div(id="out")]
-    )
+    app.layout = html.Div([html.Div("child", id="input-output"), html.Div(id="out")])
 
     with pytest.raises(CallbackException) as err:
 
         @app.callback(
-            Output("input-output", "children"),
-            [Input("input-output", "children")],
+            Output("input-output", "children"), [Input("input-output", "children")],
         )
         def failure(children):
             pass
@@ -844,11 +810,7 @@ def test_inin018_output_input_invalid_callback():
 def test_inin019_callback_dep_types():
     app = Dash(__name__)
     app.layout = html.Div(
-        [
-            html.Div("child", id="in"),
-            html.Div("state", id="state"),
-            html.Div(id="out"),
-        ]
+        [html.Div("child", id="in"), html.Div("state", id="state"), html.Div(id="out")]
     )
 
     with pytest.raises(IncorrectTypeException):
@@ -911,8 +873,7 @@ def test_inin020_callback_return_validation():
         pytest.fail("not serializable")
 
     @app.callback(
-        [Output("c", "children"), Output("d", "children")],
-        [Input("a", "children")],
+        [Output("c", "children"), Output("d", "children")], [Input("a", "children")],
     )
     def multi(a):
         return [1, set([2])]
@@ -922,8 +883,7 @@ def test_inin020_callback_return_validation():
         pytest.fail("nested non-serializable")
 
     @app.callback(
-        [Output("e", "children"), Output("f", "children")],
-        [Input("a", "children")],
+        [Output("e", "children"), Output("f", "children")], [Input("a", "children")],
     )
     def multi2(a):
         return ["abc"]
@@ -939,15 +899,10 @@ def test_inin021_callback_context(dash_duo):
     btns = ["btn-{}".format(x) for x in range(1, 6)]
 
     app.layout = html.Div(
-        [
-            html.Div([html.Button(btn, id=btn) for btn in btns]),
-            html.Div(id="output"),
-        ]
+        [html.Div([html.Button(btn, id=btn) for btn in btns]), html.Div(id="output")]
     )
 
-    @app.callback(
-        Output("output", "children"), [Input(x, "n_clicks") for x in btns]
-    )
+    @app.callback(Output("output", "children"), [Input(x, "n_clicks") for x in btns])
     def on_click(*args):
         if not callback_context.triggered:
             raise PreventUpdate
@@ -1019,3 +974,11 @@ def test_inin023_wrong_callback_id():
     @app.callback(Output("inner-div", "children"), [Input("inner-input", "value")])
     def h(a):
         return a
+
+
+def test_inin_024_port_env_success(dash_duo):
+    app = Dash(__name__)
+    app.layout = html.Div("hi", "out")
+    dash_duo.start_server(app, port="12345")
+    assert dash_duo.server_url == "http://localhost:12345"
+    dash_duo.wait_for_text_to_equal("#out", "hi")